+++ Cuckoofile.toml
@@ -128,6 +128,9 @@ sources = [
     "novawallet/Common/Services/RemoteSubscription/WalletRemoteSubscription/WalletRemoteSubscriptionFactory.swift",
     "novawallet/Common/DataProvider/GeneralStorageSubscriptionFactory.swift",
     "novawallet/Common/DataProvider/GiftsLocalSubscriptionFactory.swift",
+    "novawallet/Common/Substrate/AssetOperations/AssetStorageInfoOperationFactory.swift",
+    "novawallet/Common/Services/RemoteSubscription/Substrate/WalletRemoteQueryWrapperFactory.swift",
+    "novawallet/Common/Services/AssetExchange/AssetsExchangeGraph.swift",
 ]
 
 exclude = [
